refactor: move notification factories under Observable.ts
- This is in preparation for moving Observable and related types to its own packages
