+++ packages/rxjs/spec/operators/dematerialize-spec.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import { of, ObservableNotification, Observable } from 'rxjs';
-import { COMPLETE_NOTIFICATION, errorNotification, nextNotification } from 'rxjs/internal/NotificationFactories';
+import { COMPLETE_NOTIFICATION, errorNotification, nextNotification } from 'rxjs/internal/Observable';
 import { dematerialize, map, mergeMap, materialize, take } from 'rxjs/operators';
 import { TestScheduler } from 'rxjs/testing';
 import { observableMatcher } from '../helpers/observableMatcher';
