+++ packages/rxjs/spec/schedulers/TestScheduler-spec.ts
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import { TestScheduler } from 'rxjs/testing';
 import { Observable, Subject, of, merge, animationFrameScheduler, asapScheduler, asyncScheduler, interval } from 'rxjs';
 import { delay, debounceTime, concatMap, mergeMap, mapTo, take } from 'rxjs/operators';
-import { nextNotification, COMPLETE_NOTIFICATION, errorNotification } from 'rxjs/internal/NotificationFactories';
+import { nextNotification, COMPLETE_NOTIFICATION, errorNotification } from 'rxjs/internal/Observable';
 import { animationFrameProvider } from 'rxjs/internal/scheduler/animationFrameProvider';
 import { immediateProvider } from 'rxjs/internal/scheduler/immediateProvider';
 import { intervalProvider } from 'rxjs/internal/scheduler/intervalProvider';
